Fixed level requirements not loading correctly and delete some useless tests on levels.
